oliwia_agent_azure/pyproject.toml
[project]
name = "a2a-oliwia-agent-azure"
version = "1.0.0"
description = "Oliwia's Azure OpenAI scheduling agent with A2A Protocol"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.5,<0.2.7",
    "httpx>=0.28.1",
    "openai>=1.50.0",
    "openai-agents>=0.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
